+++ doc/source/conf.py
@@ -1,3 +1,8 @@
+import os
+import re
+import sys
+import mplcursors
+
 # -- General configuration ------------------------------------------------
 
 extensions = [
@@ -20,8 +25,6 @@ copyright = '2016–present, Antony Lee'
 author = 'Antony Lee'
 
 # RTD modifies conf.py, making versioneer mark the version as -dirty.
-import re
-import mplcursors
 version = release = re.sub(r'\.dirty$', '', mplcursors.__version__)
 
 language = 'en'
@@ -41,7 +44,9 @@ html_theme_options = {
     'github_user': 'anntzer',
     'github_repo': 'mplcursors',
     'github_banner': True,
-    'github_button': False}
+    'github_button': False,
+    'code_font_size': '80%',
+}
 # html_last_updated_fmt = ''  # bitprophet/alabaster#93
 
 htmlhelp_basename = 'mplcursors_doc'
@@ -49,25 +54,35 @@ htmlhelp_basename = 'mplcursors_doc'
 # -- Options for LaTeX output ---------------------------------------------
 
 latex_elements = {}
-latex_documents = [
-    (master_doc, 'mplcursors.tex', 'mplcursors Documentation',
-     'Antony Lee', 'manual'),
-]
+latex_documents = [(
+    master_doc,
+    'mplcursors.tex',
+    'mplcursors Documentation',
+    'Antony Lee',
+    'manual',
+)]
 
 # -- Options for manual page output ---------------------------------------
 
-man_pages = [
-    (master_doc, 'mplcursors', 'mplcursors Documentation',
-     [author], 1)
-]
+man_pages = [(
+    master_doc,
+    'mplcursors',
+    'mplcursors Documentation',
+    [author],
+    1,
+)]
 
 # -- Options for Texinfo output -------------------------------------------
 
-texinfo_documents = [
-    (master_doc, 'mplcursors', 'mplcursors Documentation',
-     author, 'mplcursors', 'Interactive data selection cursors for Matplotlib.',
-     'Miscellaneous'),
-]
+texinfo_documents = [(
+    master_doc,
+    'mplcursors',
+    'mplcursors Documentation',
+    author,
+    'mplcursors',
+    'Interactive data selection cursors for Matplotlib.',
+    'Miscellaneous',
+)]
 
 # -- Misc. configuration --------------------------------------------------
 
@@ -76,17 +91,20 @@ autodoc_member_order = 'bysource'
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3', None),
     'matplotlib': ('https://matplotlib.org', None),
-    'pandas': ('http://pandas.pydata.org/pandas-docs/stable', None)}
+    'pandas': ('http://pandas.pydata.org/pandas-docs/stable', None),
+}
 
 # CustomSortKey cannot be defined *here* because it would be unpicklable as
 # this file is exec'd rather than imported.
-import sys; sys.path.append(".")
+sys.path.append(".")
 from _local_ext import CustomSortKey
+
+os.environ.pop("DISPLAY", None)  # Don't warn about non-GUI when running s-g.
 
 sphinx_gallery_conf = {
     'backreferences_dir': False,
     'examples_dirs': '../../examples',
-    'filename_pattern': '.*\.py',
+    'filename_pattern': r'.*\.py',
     'gallery_dirs': 'examples',
     'min_reported_time': 1,
     'within_subsection_order': CustomSortKey,